feat(parse): manual parsing

+++ compiler/ac_ir/src/span.rs
@@ -3,9 +3,9 @@ use crate::source::FileId;
 #[salsa::tracked(debug)]
 pub struct Span<'db> {
     #[tracked]
-    pub lo: usize,
+    pub lo: u32,
     #[tracked]
-    pub hi: usize,
+    pub hi: u32,
     #[tracked]
     pub file: FileId<'db>,
 }
